requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
PyYAML==6.0.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
smmap==5.0.1
urllib3==2.2.2
